moved service mfs_activity to service/private
